requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
telethon==1.37.0
pymongo==4.9.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0